+++ morpheus-graphql-core/changelog.md
@@ -10,6 +10,8 @@
 
 ### Breaking changes
 
+- removed `parseDSL` from `Data.Morpheus.Core`
+
 - following Types and modules are migrated to the new package `morpheus-graphql-app`:
 
   - following types and functions in `Data.Morpheus.Core` are moved in to `Data.Morpheus.App`: